% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camcog.R
\docType{data}
\name{camcog}
\alias{camcog}
\title{CAMCOG data}
\format{
Table with 2 columns and 136 rows:
\describe{
\item{diag_group}{Cognitive impairment groups: control (CG),
mild cognitive impairment (MCI) and Alzheimer's disease (AD)}
\item{CAMCOG}{CAMCOG score}
}
}
\source{
J.F. Cecato, J.E. Martinelli, R. Izbicki, M.S. Yassuda, and
I. Aprahamian, 2016: A subtest analysis of the montreal cognitive assessment
(moca): which subtests can best discriminate between healthy controls, mild
cognitive impairment and alzheimer’s disease? International psychogeriatrics,
28, no. 5, 825–832.
}
\usage{
camcog
}
\description{
Data from Cecato et al. (2016) consisting of an application of the Cambridge
Cognition Examination (CAMCOG), a widely-used questionnaire for measuring the
extent of dementia and assessing the level of cognitive impairment.
}
\keyword{datasets}
