% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_comparisons.R
\name{m_comparisons}
\alias{m_comparisons}
\title{REACT for multiple parameters comparisons with logical consistency}
\usage{
m_comparisons(alpha = 0.05, nrow = NA, ncol = NA, delta, par, f_matrix)
}
\arguments{
\item{alpha}{alpha level (default is 0.05)}

\item{nrow}{Number of rows for multiple plotting. If NA, we choose based on vector length
(default is NA)}

\item{ncol}{Number of columns for multiple plotting. If NA, we choose based on vector length
(default is NA)}

\item{delta}{delta to build pragmatics.}

\item{par}{vector of point estimates of each parameter}

\item{f_matrix}{fisher information matrix associated to the vector of parameters}
}
\description{
REACT plotting for multiple parameters comparisons. It uses asymptotic properties from
MLE to plot pairwise confidence ellipsis and conclude if we reject, accept or remain agnostic about
differences $|\theta_i - \theta_j|$ between each pair of parameter $\theta_i$ and $\theta_j$.
Returns a plot list with all plottings of each comparisson.
}
