% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNT_tests.R
\name{NNT_indep_test}
\alias{NNT_indep_test}
\title{REACT for difference between success proportions of two independent populations}
\usage{
NNT_indep_test(NNT, event_e, n_e, event_c, n_c, alpha = 0.05, verbose = TRUE)
}
\arguments{
\item{NNT}{Number Necessary to Treat used to build pragmatic hypothesis.}

\item{event_e}{Number of successes in treatment group.}

\item{n_e}{Sample size of treatment group.}

\item{event_c}{Number of successes in control (or other) group.}

\item{n_c}{Sample size of control (or other) group.}

\item{alpha}{alpha level. Default is 0.05.}

\item{verbose}{set whether text output should be generated (verbose = TRUE) or not (verbose = FALSE).
Default is TRUE.}
}
\value{
"simple REACT" object with the original confidence interval, pragmatic region derived by the chosen NNT and the test outcome
}
\description{
NNT based REACT for difference between success proportions of two independent populations.
}
