% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_test.R
\name{plot.simple_REACT}
\alias{plot.simple_REACT}
\title{Plot Region of equivalence and confidence interval of precise hypothesis test}
\usage{
\method{plot}{simple_REACT}(obj)
}
\arguments{
\item{obj}{simple_REACT object to plot}
}
\description{
Plot Region of equivalence and confidence interval of precise hypothesis test
}
