% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REACT_glm.R
\name{REACT.glm}
\alias{REACT.glm}
\title{REACT GLM covariate selection}
\usage{
REACT.glm(
  formula,
  family = gaussian,
  data,
  weights,
  subset,
  tol,
  na.action,
  alpha = 0.05,
  verbose = FALSE,
  start = NULL,
  etastart,
  mustart,
  offset,
  control = list(...),
  model = TRUE,
  method = "glm.fit",
  x = FALSE,
  y = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{tol}{Tolerance to build pragmatic hypothesis. It can be also an asymmetric tolerance (vector of length 2)}

\item{alpha}{Chosen significance level. Default is 0.05}

\item{verbose}{Set whether text output should be generated (verbose = TRUE) or not (verbose = FALSE)}
}
\value{
An object of class "simple_REACT" with the original confidence interval, pragmatic region and the test outcome. There is also a plot method available for this kind of object.
}
\description{
REACT-based feature selection of GLM covariates.
}
