% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\docType{data}
\name{react}
\alias{react}
\title{Reactivity helper}
\format{
An object of class \code{react} of length 0.
}
\usage{
react
}
\description{
The \code{react} object allow you to use the \code{react$foo} syntax
instead of calling \code{foo()}.
}
\details{
The benefit is that it makes it easier to identify calls to reactive
expressions and take them apart from calls to functions.
}
\keyword{datasets}
