% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{REACT_forestplot}
\alias{REACT_forestplot}
\title{REACT forestplot}
\usage{
REACT_forestplot(
  CI_matrix,
  NNT,
  hip = 0,
  point_estim = NULL,
  study_names = NULL
)
}
\arguments{
\item{CI_matrix}{Matrix of lower and upper bounds of confidence intervals related to each study
with columns named CIlower and CIupper}

\item{NNT}{Number Necessary to Treat used to build pragmatic hypothesis}

\item{hip}{precise hypothesis of interest - default is 0}

\item{point_estim}{Vector of point estimations related each study parameter. Default is NULL}

\item{study_names}{Vector of study names. Default is NULL}
}
\description{
REACT adapted forestplot for agnostic and pragmatic meta analysis data
}
