% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR_utils.R
\name{REACT_RRplot}
\alias{REACT_RRplot}
\title{REACT Relative Risk forestplot.}
\usage{
REACT_RRplot(
  CI_matrix,
  NNT = NULL,
  CER = 1,
  RRR = NULL,
  hyp = 1,
  point_estim = FALSE,
  covar_names = NULL
)
}
\arguments{
\item{CI_matrix}{Matrix of lower and upper bounds of confidence intervals related to each covariate
with columns named CIlower and CIupper.}

\item{NNT}{Number Necessary to Treat used to build pragmatic hypothesis. Default is NULL.}

\item{CER}{Control Error Rate used to build pragmatic hypothesis. Default is 1.}

\item{RRR}{Relative Risk Reduction used to build pragmatic hypothesis. Default is NULL.}

\item{hyp}{precise hypothesis of interest - default is 1.}

\item{point_estim}{Logical or vector of point estimations related to each covariate. If TRUE,
it uses the mean of the confidence intervals as the point estimate. Default is FALSE.}

\item{covar_names}{Vector of covariate names. Default is NULL.}
}
\value{
REACT forestplot ggplot object.
}
\description{
REACT adapted forestplot for agnostic and pragmatic meta analysis of relative risk.
}
