% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_test.R
\name{simple_test}
\alias{simple_test}
\title{REACT function for uni parametric precise hypothesis testing}
\usage{
simple_test(CI, tol, hip = 0, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{CI}{parameter confidence interval for testing}

\item{tol}{tolerance to build pragmatic hypothesis}

\item{hip}{precise hypothesis of interest - default is 0}

\item{plot}{set whether the test results should be plotted (plot = TRUE) or not (plot = FALSE), default is plot = TRUE}

\item{verbose}{set whether text output should be generated (verbose = TRUE) or not (verbose = FALSE)}
}
\description{
REACT function for uni parametric precise hypothesis testing
}
