% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_test.R
\name{base_test}
\alias{base_test}
\title{Uni-parametric Agnostic and Pragmatic Confidence-based precise test}
\usage{
base_test(CI, tol, hyp = 0, verbose = TRUE)
}
\arguments{
\item{CI}{Parameter confidence interval for testing}

\item{tol}{Tolerance to build pragmatic hypothesis. It can be also an asymmetric tolerance (vector of length 2)}

\item{hyp}{Precise hypothesis of interest - default is 0}

\item{verbose}{Set whether text output should be generated (verbose = TRUE) or not (verbose = FALSE)}
}
\value{
An object of class "simple_REACT" with the original confidence interval, pragmatic region and the test outcome. There is also a plot method available for this kind of object.
}
\description{
REACT function for uni parametric precise hypothesis testing.
}
\examples{
# chosen CI, tolerance and precise hypothesis
ci <- c(-1, 1)
tol <- 0.5
hyp <- 0
# printing
test <- base_test(ci, tol, hyp)

}
