% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_comparisons.R
\name{m_comparisons}
\alias{m_comparisons}
\title{REACT for multiple parameters comparisons with logical consistency}
\usage{
m_comparisons(
  alpha = 0.05,
  nrow = NA,
  ncol = NA,
  invert = FALSE,
  tol,
  par,
  f_matrix
)
}
\arguments{
\item{alpha}{alpha level (default is 0.05)}

\item{nrow}{Number of rows for multiple plotting. If NA, we choose based on vector length
(default is NA)}

\item{ncol}{Number of columns for multiple plotting. If NA, we choose based on vector length
(default is NA)}

\item{invert}{Boolean indicating whether fisher matrix needs to be inverted or no.
Default is FALSE}

\item{tol}{tolerance to build pragmatics.}

\item{par}{vector of point estimates of each parameter}

\item{f_matrix}{fisher information matrix associated to the vector of parameters}
}
\value{
A list with each parameter comparisons ggplot-based plots
}
\description{
REACT plotting for multiple parameters comparisons. It uses asymptotic properties from
MLE to plot pairwise confidence ellipsis and conclude if we reject, accept or remain agnostic about
each differences \eqn{|\theta_i - \theta_j|} between each pair of parameter \eqn{\theta_i} and \eqn{\theta_j}.
Returns a plot list with all plottings of each comparisson.
}
