% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\docType{data}
\name{react}
\alias{react}
\title{Reactivity helper}
\format{
An object of class \code{react} of length 0.
}
\usage{
react
}
\description{
The \code{react} object gives alternative syntax to call
shiny reactive expressions.
}
\details{
The benefit is that it makes them easier to spot in your code.
}
\examples{
\dontrun{
  # imagine you have some shiny reactive
  foo <- reactive({
    whatever() + something()
  })

  # react allows you to call it in 3 different ways
  # instead of the usual foo() form
  react$foo
  react[foo]
  react[foo()]
}
}
\keyword{datasets}
