% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{REACT_forestplot}
\alias{REACT_forestplot}
\title{REACT forestplot.}
\usage{
REACT_forestplot(
  CI_matrix,
  NNT,
  hyp = 0,
  point_estim = FALSE,
  study_names = NULL,
  fail = FALSE
)
}
\arguments{
\item{CI_matrix}{Matrix of lower and upper bounds of confidence intervals related to each study
with columns named CIlower and CIupper.}

\item{NNT}{Number Necessary to Treat used to build pragmatic hypothesis.}

\item{hyp}{precise hypothesis of interest. Default is 0.}

\item{point_estim}{Logical or vector of point estimations related to each study parameter. If TRUE,
it uses the mean of the confidence intervals as the point estimate. Default is FALSE.}

\item{study_names}{Vector of study names. Default is NULL.}

\item{fail}{Boolean specifying whether the probabilities are for failure (TRUE) or success (FALSE).
Default is FALSE.}
}
\value{
React forestplot ggplot object
}
\description{
REACT adapted forestplot for agnostic and pragmatic meta analysis data.
}
